+++ server/price-action/lib/queue/repeat-queue.ts
@@ -1,7 +1,7 @@
 import { Job, Queue, QueueEvents, QueueScheduler, Worker } from "bullmq";
 import dayjs from "dayjs";
 import Redis from "ioredis";
-import { fetchAndInsertSnapshot } from "../../polygon/requests/snapshot/insert";
+import { fetchAndInsertSnapshot } from "../polygon/requests/snapshot/insert";
 
 // BullMQ wants to connect to Redis separately. For now, we'll use the same
 // store we use elsewhere, but we might want to create a separate Docker service
